requirements.txt
requests==2.32.3
python-dotenv==1.1.0
mysql-connector-python==9.3.0
black==25.1.0
pylint==3.3.4